Environmental variables for dev/production mode (#1455)
* Restore ELECTRON_IS_DEV environment variable

As part of migrating to @electron/remote from electron.remote, 6b3fe94d7
removed the electron-is-dev package and with it the support of selecting
dev/production mode with the ELECTRON_IS_DEV environmental variable.

This commit restores support for this variable. Because even the newest
version of the electron-is-dev package breaks in renderer processes, we
instead query the environment ourselves.

* Add support for NODE_ENV variable

Also support NODE_ENV for specifying dev mode in addition to the
ELECTRON_IS_DEV variable. This variable is used by e.g., the packaging
in Arch Linux to trigger production mode with an explicit electron
command line invocation:
https://aur.archlinux.org/cgit/aur.git/tree/PKGBUILD?h=ferdi-git&id=61dc59e5eb19a2c9e8f9edaf0a63aaae72990a0b#n109

* Refactor environmental variable handling