fix: modify user agent string on Linux (#99)
On Linux, the user agent includes 'X11; Ubuntu; Linux ${arch}'. This
user agent string break the Google Calendar service because the server
decides to serve to mobile site instead of the desktop site based on
that string.

On my non-Ubuntu Linux machine, the UA used is almost identical, except
for the 'Ubuntu; ' part. Removing that part from the default UA we use
fixes that particular error and seems no to break any other services.